requirements.txt
llama-index
llama-index-readers-web
langchain
langchain-community
langchain-google-community
openai
requests
python-dotenv
google-api-python-client
beautifulsoup4
lxml
numpy==1.26.4